ACDEVOPS-12 New organization scheme, preparation for new feature files, moved set up and teardown back into independent file
