+++ src/test/java/sakai/steps/LoginStep.java
@@ -1,39 +1,16 @@
 package sakai.steps;
 
-import cucumber.api.Scenario;
-import cucumber.api.java.After;
-import cucumber.api.java.Before;
 import cucumber.api.java.en.And;
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
-import io.github.bonigarcia.wdm.FirefoxDriverManager;
-import io.github.bonigarcia.wdm.ChromeDriverManager;
-import org.apache.commons.io.FileUtils;
-import org.openqa.selenium.OutputType;
-import org.openqa.selenium.TakesScreenshot;
-import org.openqa.selenium.firefox.FirefoxBinary;
-import org.openqa.selenium.firefox.FirefoxOptions;
-import sakai.utilities.Configuration;
-import sakai.utilities.SakaiLogger;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
 import sakai.pages.BasePage;
 import sakai.pages.HomePage;
 import sakai.pages.LoginPage;
-import sakai.utilities.PageWaiter;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.Date;
-
-import static org.junit.Assert.fail;
+import sakai.utilities.Util;
 
 public class LoginStep {
 
-    private WebDriver driver;
     private BasePage page;
     private LoginPage login;
     private HomePage home;
@@ -41,7 +18,7 @@ public class LoginStep {
 
     @Given("^I navigate to the login page$")
     public void iNavigateToTheLoginPage() {
-        page = new BasePage(driver);
+        page = new BasePage(Util.getDriver());
         login = page.navigateToLogin();
     }
 
@@ -50,10 +27,10 @@ public class LoginStep {
         home = login.loginAsStudent();
     }
 
-    @And("^I close What's New Popup if needed$")
+    @And("^I close new feature popup if needed$")
     public void iCloseWhatSNewPopupIfNeeded() { home.closeNewFeaturePopUp();}
 
-    @Then("^I should see NYU Classes logo$")
+    @Then("^I should see Sakai logo$")
     public void iShouldSeeNYUClassesLogo() {
         home.checkForSakaiBanner();
     }
@@ -70,77 +47,7 @@ public class LoginStep {
     }
 
     @Then("^I should see logged out$")
-    public void iShouldSeeLoggedOut() throws Throwable {
+    public void iShouldSeeLoggedOut() {
         login.checkForLoggedOutBanner();
-    }
-
-    @Before
-    public void startUp(Scenario scenario)
-    {
-        SakaiLogger.logInfo("=========== " + scenario.getName() + " ===========");
-        Configuration.setPlatform(System.getenv("sakai_browser"));
-
-        if(Configuration.getPlatform() != null && Configuration.getPlatform().equalsIgnoreCase("chrome"))
-        {
-            SakaiLogger.logInfo("Initializing testing environment");
-            SakaiLogger.logInfo("Using driver: Chrome");
-            ChromeDriverManager.getInstance().setup();
-            ChromeOptions options = new ChromeOptions();
-            options.addArguments("headless");
-            driver = new ChromeDriver(options);
-            driver.manage().deleteAllCookies();
-            PageWaiter.setDriver(driver);
-        }
-        else if(Configuration.getPlatform() != null && Configuration.getPlatform().equalsIgnoreCase("firefox"))
-        {
-            SakaiLogger.logInfo("Initializing testing environment");
-            SakaiLogger.logInfo("Using driver: Firefox");
-            FirefoxDriverManager.getInstance().setup();
-            FirefoxBinary binary = new FirefoxBinary();
-            binary.addCommandLineOptions("--headless");
-            FirefoxOptions options = new FirefoxOptions();
-            options.setBinary(binary);
-            driver = new FirefoxDriver(options);
-            driver.manage().deleteAllCookies();
-            PageWaiter.setDriver(driver);
-        }
-        else
-        {
-            SakaiLogger.logInfo("No compatible browser environment found, your browser specification is: " + Configuration.getPlatform());
-            fail("No compatible browser environment found");
-        }
-    }
-
-    @After
-    public void tearDown(Scenario scenario)
-    {
-        if(Configuration.getPlatform() == null || (!Configuration.getPlatform().equalsIgnoreCase("chrome") && !Configuration.getPlatform().equalsIgnoreCase("firefox")))
-        {
-            SakaiLogger.logInfo("Scenario failed because no browser environment was defined");
-            SakaiLogger.logInfo("=========== " + scenario.getName() + " ===========");
-        }
-        else if(scenario.isFailed())
-        {
-            SakaiLogger.logErr("Scenario failed =( - (" + scenario.getName() + ")");
-            File screencap = ((TakesScreenshot) driver).getScreenshotAs(OutputType.FILE);
-            try {
-                FileUtils.copyFile(screencap, new File("target/screencap/" + scenario.getName() + "_" + new Date() + ".png"));
-                SakaiLogger.logInfo("Cleaning the environment");
-                driver.manage().deleteAllCookies();
-                driver.quit();
-                SakaiLogger.logInfo("=========== " + scenario.getName() + " ===========");
-            }
-            catch(IOException e)
-            {
-
-            }
-        }
-        else
-        {
-            SakaiLogger.logInfo("Cleaning the environment");
-            driver.manage().deleteAllCookies();
-            driver.quit();
-            SakaiLogger.logInfo("=========== " + scenario.getName() + " ===========");
-        }
     }
 }
