--- src/test/java/sakai/utilities/Configuration.java
@@ -1,15 +1,0 @@
-package sakai.utilities;
-
-public class Configuration {
-
-    private static String platform;
-
-    public static void setPlatform(String platform){
-        Configuration.platform = platform;
-    }
-
-    public static String getPlatform()
-    {
-        return Configuration.platform;
-    }
-}
